Add support to handle 4x4 keypad
Added support to handle 4x4 keypad, using 4 GPIO pins as
outputs and 4 as inputs, and constantly polling the lines
to see if a high input causes a high output, indicating a
button has been pressed. The particular button is determined
by which pair of input and output lines are high.